支持设置是否在启动后删除日志文件

+++ app/src/main/cpp/core.cpp
@@ -24,6 +24,7 @@ static std::map<uint32_t, time_t> toastedApplication;
 static std::map<uint32_t, time_t> ignoredProcess;
 static short packageSearchDepth = 1;
 static bool checkSuCompat = false;
+static bool autoDeleteLog= false;
 
 struct __attribute__((packed)) EventRecordHeader {
     uint16_t record_type;
@@ -164,7 +165,7 @@ bool handleSuLog() {
     }
     std::thread pollingThread(pollingLogEvent, suLogFd);
     pollingThread.detach();
-    deleteSuLogFile();
+    if (autoDeleteLog) deleteSuLogFile();
     return true;
 }
 
@@ -186,9 +187,10 @@ JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM *vm, void *reserved) {
 extern "C"
 JNIEXPORT jboolean JNICALL
 Java_com_suisho_kernelsugranttoast_Entry_jniInit(JNIEnv *env, jclass clazz, short searchDepth,
-                                                 jboolean checkCompat) {
+                                                 jboolean checkCompat,jboolean deleteLog) {
     packageSearchDepth = searchDepth;
     checkSuCompat = checkCompat;
+    autoDeleteLog = deleteLog;
     if (!utilInit()) return false;
     if (!handleSuLog()) return false;
     LOGI("JNI utilInit successful");
